refracrtor: user switch
